Naive Layer Structure and Api connection

+++ app/src/main/java/com/r2ufuk/popgoesmyact/domain/entities/Actor.java
@@ -2,11 +2,11 @@ package com.r2ufuk.popgoesmyact.domain.entities;
 
 public class Actor {
     private final int id;
-    private final float popularity;
+    private final double popularity;
     private final String name;
     private final String profile_path;
 
-    public Actor(int id, float popularity, String name, String profile_path) {
+    public Actor(int id, double popularity, String name, String profile_path) {
         this.id = id;
         this.name = name;
         this.popularity = popularity;
